+++ rollup.config.ts
@@ -39,6 +39,7 @@ export default [
       "@kodepandai/flydrive",
       "@kodepandai/flydrive-s3",
       "@kodepandai/node-input-validator",
+      "@kodepandai/hash-equals",
       "formidable",
       "commander",
       "colorette",
@@ -48,6 +49,8 @@ export default [
       "objection",
       "pluralize",
       "express-session",
+      "cookie",
+      "repl",
     ],
   },
   {