+++ src/Console/Command.ts
@@ -1,8 +1,11 @@
 import { spawn } from "child_process";
 import { bgRed, cyanBright, whiteBright, yellow } from "colorette";
+import type { Application } from "../Foundation";
 import type { ObjectOf } from "../Types";
 
 class Command {
+  protected lunox!: Application;
+
   SUCCESS: 0 = 0;
   FAILURE: 1 = 1;
   INVALID: 2 = 2;
@@ -18,7 +21,7 @@ class Command {
   }
 
   public getSignature() {
-    return this.signature;
+    return this.signature.replace(/\s\s+/g, " ");
   }
   public getDescription() {
     return this.description;
@@ -30,6 +33,10 @@ class Command {
 
   public setOptions(opts: ObjectOf<any>) {
     this.opts = opts;
+  }
+
+  public setLunox(lunox: Application) {
+    this.lunox = lunox;
   }
 
   public arguments() {