+++ src/Console/Kernel.ts
@@ -8,10 +8,9 @@ import RegisterFacades from "../Foundation/Bootstrap/RegisterFacades";
 import RegisterProviders from "../Foundation/Bootstrap/RegisterProviders";
 import BootProviders from "../Foundation/Bootstrap/BootProviders";
 import fs from "fs";
-import path from "path";
 import type Command from "./Command";
 import { Command as CommanderCommand } from "commander";
-import { blue } from "colorette";
+import { bgRed, blue, whiteBright } from "colorette";
 import { exit } from "process";
 import MakeMigrationCommand from "./MakeMigrationCommand";
 import type { ObjectOf } from "../Types";
@@ -26,6 +25,9 @@ import MakeCommand from "./MakeCommand";
 import MakeMiddlewareCommand from "./MakeMiddlewareCommand";
 import MakeProviderCommand from "./MakeProviderCommand";
 import MakeControllerCommand from "./MakeControllerCommand";
+import { RuntimeException } from "../Foundation/Exception";
+import KeyGenerateCommand from "./KeyGenerateCommand";
+import TinkerCommand from "./TinkerCommand";
 
 class Kernel {
   protected app: Application;
@@ -44,6 +46,7 @@ class Kernel {
   constructor(app: Application) {
     this.app = app;
     this.program = new CommanderCommand();
+    process.env.APP_RUNNING_IN_CONSOLE = "true";
   }
 
   public async handle() {
@@ -69,6 +72,7 @@ class Kernel {
    */
   protected async builtinCommands() {
     const commands = [
+      KeyGenerateCommand,
       MakeCommand,
       MakeControllerCommand,
       MakeMiddlewareCommand,
@@ -81,6 +85,7 @@ class Kernel {
       RollbackMigrationCommand,
       ResetMigrationCommand,
       RefreshMigrationCommand,
+      TinkerCommand,
     ];
     await Promise.all(
       commands.map((c) => {
@@ -97,20 +102,8 @@ class Kernel {
   }
 
   protected async load(paths: string) {
-    let files: string[] = [];
-    const walkDir = async (_path: string) => {
-      const _files = fs.readdirSync(_path);
-      await Promise.all(
-        _files.map(async (f) => {
-          if (fs.lstatSync(path.join(_path, f)).isDirectory()) {
-            return walkDir(path.join(_path, f));
-          }
-          files = files.concat(path.join(_path, f));
-        })
-      );
-    };
     // resolve all commands from given path
-    walkDir(paths);
+    const files = await walkDir(paths);
     // register all commands to artisan
     await Promise.all(
       files.map(async (f) => {
@@ -133,14 +126,30 @@ class Kernel {
           .filter((a) => !(a.startsWith("--") || a.startsWith("-")))
           .map((a) => a.replace("?", ""));
         const inputArgs = _program.args.reduce((p, c, i) => {
-          p[argKeys[i].split(" : ")[0]] = c;
+          if (argKeys.length > 0) {
+            p[argKeys[i].split(" : ")[0]] = c;
+          }
           return p;
         }, {} as ObjectOf<string>);
 
         commandInstance.setArguments(inputArgs);
         commandInstance.setOptions(_program.opts());
-        const exitCode = await commandInstance.handle();
-        exit(exitCode);
+        commandInstance.setLunox(this.app);
+        try {
+          const exitCode = await commandInstance.handle();
+          if (exitCode > 0) {
+            exit(exitCode);
+          }
+        } catch (error) {
+          if (error instanceof Error) {
+            console.log(bgRed(whiteBright(error.message)));
+          } else {
+            console.log(error);
+          }
+          if (!(error instanceof RuntimeException)) {
+            exit(1);
+          }
+        }
       });
 
     // parse arguments and options