+++ src/Foundation/Exception/RuntimeException.ts
@@ -1,0 +1,6 @@
+class RuntimeException extends Error {
+  constructor(message: string) {
+    super(message);
+  }
+}
+export default RuntimeException;