+++ src/Foundation/Http/Kernel.ts
@@ -1,13 +1,24 @@
 import { STATUS_CODES } from "http";
 import sirv from "sirv";
 import path from "path";
-import polka, { Request, NextHandler, Response, IError } from "polka";
+import polka, {
+  Request as ServerRequest,
+  NextHandler,
+  Response as ServerResponse,
+  IError,
+} from "polka";
 
 import type { Bootstrapper } from "../../Contracts/Foundation/Boostrapper";
-import type { Middleware } from "../../Contracts/Http/Middleware";
+import type {
+  Middleware,
+  NativeMiddleware,
+} from "../../Contracts/Http/Middleware";
 import HttpRequest from "../../Http/Request";
 import HttpResponse from "../../Http/Response";
-import Route from "../../Support/Facades/Route";
+import {
+  Route,
+  Response,
+} from "../../Support/Facades";
 import type { Class, ObjectOf } from "../../Types";
 import type Application from "../Application";
 import BootProviders from "../Bootstrap/BootProviders";
@@ -25,9 +36,9 @@ import NotFoundHttpException from "../../Http/NotFoundHttpException";
 
 class Kernel {
   protected app: Application;
-  protected middleware: Middleware[] = [];
-  protected middlewareGroups: ObjectOf<Middleware[]> = {};
-  protected routeMiddleware: ObjectOf<Middleware> = {};
+  protected middleware: (Middleware | Class<Middleware>)[] = [];
+  protected middlewareGroups: ObjectOf<(Middleware | Class<Middleware>)[]> = {};
+  protected routeMiddleware: ObjectOf<Middleware | Class<Middleware>> = {};
 
   protected bootstrappers: Class<Bootstrapper>[] = [
     LoadEnvirontmentVariabel,
@@ -54,7 +65,7 @@ class Kernel {
           res,
           response.getStatus(),
           response.getOriginal(),
-          response.getHeaders()
+          response.headers
         );
       },
       onNoMatch: () => {
@@ -67,10 +78,12 @@ class Kernel {
     await this.app.bootstrapWith(this.bootstrappers);
 
     server.use((req, res, next) => {
-      // create Http\Request on first middleware
+      // create Http\Request and Http\Response on first middleware
       // and inject it to rest of middleware
       const request = new HttpRequest(this.app, req);
+      const response = Response.make({}).setServerResponse(res);
       (req as any)._httpRequest = request;
+      (res as any)._httpResponse = response;
       if (req.method.toLowerCase() == "get") return next();
 
       const form = formidable({ multiples: true });
@@ -92,7 +105,7 @@ class Kernel {
     // run global middlewares
     const globalMiddlewares = this.middleware.map((middleware) =>
       this.handleMiddleware(middleware)
-    );
+    ) as any;
     server.use(...globalMiddlewares);
 
     const port = env("PORT") || 8000;
@@ -127,6 +140,7 @@ class Kernel {
           ...routeMiddlewares,
           async (req, res) => {
             const httpRequest = (req as any)._httpRequest as HttpRequest;
+            let httpResponse = (res as any)._httpResponse as HttpResponse;
             let response = await route.action(
               httpRequest,
               ...Object.values(req.params)
@@ -139,17 +153,48 @@ class Kernel {
             if (response instanceof RedirectResponse) {
               response.setRequest(httpRequest);
             }
+
+            const afterMiddlewares = route.middleware
+              .reduce((collect, middleware) => {
+                if (
+                  typeof middleware == "string" &&
+                  this.middlewareGroups[middleware]
+                ) {
+                  collect = [
+                    ...collect,
+                    ...this.middlewareGroups[middleware].map((m) =>
+                      this.handleMiddleware(m, true)
+                    ),
+                  ];
+                } else {
+                  collect = [
+                    ...collect,
+                    this.handleMiddleware(middleware, true),
+                  ];
+                }
+                return collect;
+              }, [] as any[])
+              .filter((m) => m != undefined);
+            for (let i = 0; i <= afterMiddlewares.length; i++) {
+              const afterMiddleware = afterMiddlewares[i];
+              if (typeof afterMiddleware == "function") {
+                httpResponse = await afterMiddleware(httpResponse);
+              }
+            }
+
             if (response instanceof HttpResponse) {
               // make sure all session is saved
               await httpRequest.session().save();
-
+              httpResponse.mergeResponse(response);
+              httpResponse.setCookiesToHeaders();
               return this.send(
                 res,
-                response.getStatus(),
-                response.getOriginal(),
-                response.getHeaders()
+                httpResponse.getStatus(),
+                httpResponse.getOriginal(),
+                httpResponse.headers
               );
             }
+
             if (["object", "number", "boolean"].includes(typeof response)) {
               return res.end(JSON.stringify(response));
             }
@@ -184,28 +229,53 @@ class Kernel {
     server.use(pub);
 
     server.listen(port, () => {
-      console.log("server run on port: " + port);
+      if (process.env.NODE_ENV != "production") {
+        return console.log(
+          "Starting development server: http://localhost:" + port
+        );
+      }
+      return console.log("server run on port: " + port);
     });
   }
 
-  private handleMiddleware(middleware: string | Middleware) {
-    const { handle } = (typeof middleware == "string"
-      ? this.routeMiddleware[middleware]
-      : middleware) || { handle: null };
-    if (!handle) throw new Error("cannot resolve middleware " + middleware);
-    return (_req: Request, res: Response, next: NextHandler) => {
+  private handleMiddleware(
+    middleware: string | Middleware | Class<Middleware>,
+    after = false
+  ) {
+    let middlewareInstance: Middleware | Class<Middleware> =
+      typeof middleware == "string"
+        ? this.routeMiddleware[middleware]
+        : middleware;
+    if (is_class(middlewareInstance)) {
+      middlewareInstance = new (middlewareInstance as Class<Middleware>)();
+    }
+    if (after) {
+      return (<Middleware>middlewareInstance).handleAfter?.bind(middlewareInstance);
+    }
+    if ((<Middleware>middlewareInstance).handleNative) {
+      return (<Middleware>middlewareInstance).handleNative?.bind(middlewareInstance) as NativeMiddleware;
+    }
+    return async (
+      _req: ServerRequest,
+      _res: ServerResponse,
+      next: NextHandler
+    ) => {
       try {
-        return handle(
-          (_req as any)._httpRequest,
-          (req: HttpRequest, nativeMiddleware) => {
-            // update instance of request from middleware next function
-            (_req as any)._httpRequest = req;
-            if (nativeMiddleware) {
-              return nativeMiddleware(_req, res, next);
-            }
-            return next();
-          }
+        const handle = (<Middleware>middlewareInstance).handle?.bind(
+          middlewareInstance
         );
+        if (handle) {
+          const responseHandle = await handle(
+            (_req as any)._httpRequest,
+            (req: HttpRequest) => {
+              // update instance of request from middleware next function
+              (_req as any)._httpRequest = req;
+              return (_res as any)._httpResponse as HttpResponse;
+            }
+          );
+          (_res as any)._httpResponse  = responseHandle;
+        }
+        return next();
       } catch (error) {
         if (error instanceof Error) {
           return next(error);
@@ -215,7 +285,7 @@ class Kernel {
   }
 
   private send(
-    res: Response,
+    res: ServerResponse,
     code = 200,
     data: any = "",
     headers: ObjectOf<string> = {}
@@ -226,7 +296,9 @@ class Kernel {
     let k: any;
     const obj: ObjectOf<any> = {};
     for (k in headers) {
-      obj[k.toLowerCase()] = headers[k];
+      if(typeof headers[k]!="function"){
+        obj[k.toLowerCase()] = headers[k];
+      }
     }
 
     let type = obj[TYPE] || res.getHeader(TYPE);