+++ src/Session/SessionManager.ts
@@ -2,9 +2,10 @@ import type { ObjectOf } from "../Types";
 import type Application from "../Foundation/Application";
 import type { Store } from "express-session";
 import type Repository from "../Config/Repository";
-import type { Configuration } from "../Contracts/Session";
+import type { SessionConfig } from "../Contracts/Config";
 import type { Session } from "express-session";
 import type Request from "../Http/Request";
+import { Str } from "../Support";
 
 interface ExtendedSession extends Partial<Session> {
   __old?: any;
@@ -19,6 +20,8 @@ class SessionManager {
 
   protected request!: Request;
 
+  protected started = false;
+
   constructor(app: Application) {
     this.app = app;
     this.session = {};
@@ -27,6 +30,9 @@ class SessionManager {
   public setRequest(request: Request) {
     this.session = request.getOriginalRequest().session || {};
     this.request = request;
+    if (!this.isStarted()) {
+      this.start();
+    }
     return this;
   }
 
@@ -138,7 +144,7 @@ class SessionManager {
     }
   }
 
-  public getConfig(): Configuration {
+  public getConfig(): SessionConfig {
     return this.app.make<Repository>("config").get("session");
   }
 
@@ -165,6 +171,25 @@ class SessionManager {
         }
       });
     });
+  }
+
+  public async start() {
+    if (!this.has("_token")) {
+      this.regenerateToken();
+    }
+    this.started = true;
+  }
+
+  public regenerateToken() {
+    this.put("_token", Str.random(40));
+  }
+
+  public token() {
+    return this.get("_token");
+  }
+
+  public isStarted() {
+    return this.started;
   }
 }
 export default SessionManager;