+++ src/Support/index.ts
@@ -2,5 +2,6 @@ import ServiceProvider from "./ServiceProvider";
 import { type Trait, Traitable } from "./Traitable";
 import useMagic from "./useMagic";
 import Str from "./Str";
+import Arr from "./Collection/Arr";
 
-export { ServiceProvider, Trait, Traitable, useMagic, Str };
+export { ServiceProvider, Trait, Traitable, useMagic, Str, Arr };