+++ src/helpers.ts
@@ -9,6 +9,7 @@ import path from "path";
 import { fileURLToPath } from "url";
 import View from "./Support/Facades/View";
 import crypto from "crypto";
+import fs from "fs";
 
 declare global {
   interface Window {
@@ -29,11 +30,14 @@ declare global {
   var back: () => RedirectResponse;
   var sha1: (value: string) => string;
   var stub_path: (path?: string) => string;
+  var lunox_path: (path?: string) => string;
   var abort: (
     code: number,
     message?: string,
     headers?: ObjectOf<string>
   ) => void;
+  var is_class: (instance: any) => boolean;
+  var walkDir: (path: string) => Promise<string[]>;
 }
 
 global.get_current_dir = (importMetaUrl: string) => {
@@ -58,5 +62,26 @@ global.sha1 = (value) => crypto.createHash("sha1").update(value).digest("hex");
 global.stub_path = (_path = "") =>
   path.join(get_current_dir(import.meta.url), "..", "stub", _path);
 
+global.lunox_path = (_path = "") =>
+  path.join(get_current_dir(import.meta.url), _path);
+
 global.abort = (code: number, message = "", headers: ObjectOf<string> = {}) =>
   app().abort(code, message, headers);
+global.is_class = (instance: any) =>
+  typeof instance === "function" && /^class\s/.test(instance + "");
+
+global.walkDir = async (_path: string) => {
+  let files: string[] = [];
+  const _files = fs.readdirSync(_path);
+  await Promise.all(
+    _files.map(async (f) => {
+      if (fs.lstatSync(path.join(_path, f)).isDirectory()) {
+        const result = await walkDir(path.join(_path, f));
+        files = files.concat(result);
+        return;
+      }
+      files = files.concat(path.join(_path, f));
+    })
+  );
+  return files;
+};